+++ CHANGELOG.md
@@ -1,10 +1,22 @@
 # apollo
 
+## 0.8.0
+
+### Minor Changes
+
+- a03bbea: vocab mapper: Add batching
+
+### Patch Changes
+
+- 1d824ba: job_chat: Update the AI Assistant prompt to be more concise, secure,
+  and context-aware
+- 960dffb: job_chat: add basic prompt support for collections
+
 ## 0.7.2
 
 ### Patch Changes
 
-- ob_chat: Fix depenencies
+- ob_chat: Fix dependencies
 - vocab_mapper: Fix production key loading
 
 ## 0.7.1